refactor: 기능별 폴더 분류
